Reorganized and documented data structures for pipeline communications.
Because these new structures use sets, functions pipe_comm()
and pipe_getAndUnblock() work out of the box, so the contents have been
uncommented.

Also, the functions to register a loop and an iteration have been removed.
This is probably something that we want to do at compile time, not at
execution time.