Bugfix: change the allLocalData datatype from deque to map
In some places, the deque was accessed by index although we do not know
for sure that array X is going to be in position X of allLocalData. This
was ok (although ugly) for the old implementation where all processes
have a local version of all arrays, but this is not the case anymore.